requirements.txt
Django==3.1.7
djangorestframework==3.12.4
django-phonenumber-field==6.1.0
Babel==2.9.1
